@@ -1,11 +1,12 @@
 package local.concept2_vc;
 //test para detectar los cornerCase (casos raros)
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertTrue;
-import static org.junit.Assert.fail;
 
-import org.junit.Before;
-import org.junit.Test;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.junit.jupiter.api.Assertions.fail;
+
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 
 import local.exceptions.BusinessException;
 import local.exceptions.ErrorCodes;
@@ -14,7 +15,7 @@ public class CalcControllerTest {
     
     private CalcController calculator;
     
-    @Before
+    @BeforeEach
     public void config(){
         calculator = new CalcController();
     }
